requirements.txt
pydantic
fastapi
pydantic_ai
langchain
langchain_openai
python-dotenv
firecrawl
praw
pytrends
requests
uvicorn
websockets
uuid
openai-agents